feat: add ShowAlphaSlider/Vertical to ColorTriangle
Restructures ColorTriangle as a composite mirroring the ColorWheel /
ColorDisc pattern:

- New ColorTriangleArea (the existing SkiaPickerBase that draws the
  triangle + hue ring; renamed file ColorTriangle.cs -> ColorTriangleArea.cs)
- New ColorTriangle : ColorPickerBase composite that hosts the area
  plus an optional AlphaSlider, with the same Measure/Arrange layout
  logic ColorWheel uses (slider fraction = 0.1 per slider, slider
  thickness aspect-locked to the area size).

New BindableProperties on the composite:
  - ShowAlphaSlider (bool, default false)
  - Vertical (bool, default false)

Forwarded to the inner area:
  - RotateTriangleByHue
  - CanvasBackgroundColor
  - IndicatorRadiusScale (default 0.035F, matching the previous
    ColorTriangle constructor)

Defaults preserve existing visual behavior: ShowAlphaSlider=false ->
the area gets the full bounds, identical to today.

Public API change: the static helpers ColorToHsv / ColorFromHsv
previously on ColorTriangle now live on ColorTriangleArea. Pre-1.0,
no published consumers.

MainPage.xaml now binds ColorTriangle1.ShowAlphaSlider to the existing
ShowAlphaSwitch so the new feature is visible in the test app.
